DP/requirements.txt
absl-py==2.3.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2022.12.7
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.10.0
defusedxml==0.7.1
dnspython==2.7.0
Flask==2.3.3
flatbuffers==25.2.10
fonttools==4.58.1
gast==0.4.0
google-auth==2.40.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==2.10
itsdangerous==2.2.0

Jinja2==3.1.6
joblib==1.5.1
keras==2.13.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.7.3
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
opt_einsum==3.4.0
packaging==25.0
Pillow==9.5.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pymongo==4.6.1
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
roboflow==1.1.6
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.15.3
sentencepiece==0.2.0
setuptools==65.5.0
six==1.17.0
sounddevice==0.5.2
supervision==0.19.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0

tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.5.0
urllib3==2.4.0
Werkzeug==2.3.7
wheel==0.45.1
wrapt==1.17.2
